Returns the payload length during data_match

+++ src/data.h
@@ -59,7 +59,7 @@
 
     data_payload_t data_get(data_root_t *root, size_t offset, size_t length, uint16_t dataid, uint8_t idlength);
     int data_check(data_root_t *root, size_t offset, uint16_t dataid);
-    int data_match(data_root_t *root, void *id, uint8_t idlength, size_t offset, uint16_t dataid);
+    size_t data_match(data_root_t *root, void *id, uint8_t idlength, size_t offset, uint16_t dataid);
 
     size_t data_insert(data_root_t *root, unsigned char *data, uint32_t datalength, void *vid, uint8_t idlength);
     size_t data_next_offset(data_root_t *root);
